Reconnect to master on non 200 heartbeat response
If heartbeat response is not HTTP 200, it means that the master is up,
but does not recognize the slave. In such cases the slave should
immediately try to reconnect to master. This event should also be
considered as heartbeat failure and increment the heartbeat failure
count. If the slave gets more failures than the threshold, the process
dies